OGC Refactoring Part 1 - Combine OgcRecordApi with OgcRecordApiController
